javapns branch 1.7
This is an in-development branch featuring significant code refactoring and reengineering intended to make the library usable by large projects.

The main goal of this branch is to do away with the previously-used singleton pattern for PushNotificationManager and DeviceFactory classes, as that pattern seriously impairs the library's usability in dependency-injected projects, and/or projects that need to persist device information.